pyproject.toml
[project]
name = "figure-webnav"
version = "0.1.0"
description = "Browser navigation challenge agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.18.0",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["webnav"]

[tool.coverage.report]
show_missing = true
fail_under = 80
